\name{testIRSN5D}
\docType{data}
\alias{testIRSN5D}
\title{A set of test data}
\description{These test data correspond to the five-dimensional case provided by the IRSN
detailed in \code{\link{dataIRSN5D}}.}
\usage{data(testIRSN5D)}
\format{A data frame with 324 rows representing the number of observations and 6 columns:
the first five corresponding to the input variables ("b","e","p","r" and "l") and the last to the response
\code{Keff}.
}
\seealso{\code{\link{dataIRSN5D}}}
\source{IRSN (Institut de Radioprotection et de Sûreté Nucléaire)}
\keyword{datasets}